app/pyproject.toml
[tool.poetry]
name = "tech_events_hub_platform"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.38.1"
pydantic-settings = "^2.7.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
psycopg = {extras = ["binary", "pool"], version = "^3.1.0"}
greenlet = "^3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
ruff = "^0.6.1"
mypy = "^1.14.1"
pytest = "*"
pytest-cov = "*"
pytest-deadfixtures = "*"
pytest-mock = "*"
pytest-repeat = "*"
pytest-asyncio = "*"
httpx = "^0.28.1"
respx = "^0.21.1"
testcontainers = {extras = ["postgres"], version = "^4.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 100
lint.select = ["ALL"]
lint.ignore = [
    "D1",
    "D203",
    "D205",
    "D213",
    "FA102",
    "ANN101",
    "ANN102",
    "TRY003",
    "EM",
    "FBT",
    "FIX",
    "TD",
    "S104",
    "FA",
    "PGH003",
    "DTZ007",
    "D401",
    "RUF001",
    "B008",
    "SIM102",
    "D404",
    "S101",
    "RET505",
]

[tool.ruff.lint.isort]
no-lines-before = ["standard-library", "local-folder"]
known-third-party = []
known-local-folder = ["whole_app"]

[tool.ruff.lint.pylint]
max-args = 8
max-branches = 8
max-bool-expr = 5
max-public-methods = 5
max-statements = 40

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLR2004", # allow magic numbers
    "SLF001",  # allow private members access
    "B017",    # allow pytest.raises(Exception)
    "PT011",   # allow pytest.raises(Exception) без match
    "S603",    # allow subprocess
    "ARG001",  # allow unused fixtures-arguments
    "ANN401",  # allow Any
]

[tool.mypy]
disable_error_code = ["import-untyped", "unused-ignore"]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_no_return = true
implicit_reexport = false
strict_equality = true


[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
